src/test/java/com/leetcode/strings/easy/ValidAnagram_242Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class ValidAnagram_242Test {

  private final ValidAnagram_242 solution = new ValidAnagram_242();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {

      return Stream.of(arguments("anagram", "nagaram", true), arguments("rat", "car", false));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void isAnagram(String str1, String str2, boolean expected) {
    boolean result = solution.isAnagram(str1, str2);

    assertThat(result).isEqualTo(expected);
  }
}
